fix: 패키지 이름 및 위치 변경
